Create .repo files pointing to dependency repos
Delorean relies on packages from other repos to add missing
dependencies, specially during heavy development when it takes time
to get packages in Fedora/EPEL repos.

Creating .repo files for each instance, so they can be
automatically fetched and used by Delorean. To put this in action,
it will require an update to the Delorean scripts.